requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bootstrap4==0.1.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-ckeditor==6.0.0
django-heroku==0.3.1
django-import-export==2.5.0
django-js-asset==1.2.2
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
news==1.0
nltk==3.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2020.11.15
pipreqs==0.4.10
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.3
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
toml==0.10.1
tornado==6.1
tqdm==4.55.0
traitlets==5.0.5
tweepy==3.10.0
urllib3==1.26.4
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
